Some clean ups to make the code compile in position independent mode.
git-svn-id: https://seccompsandbox.googlecode.com/svn/trunk@13 55e79e8e-603c-11de-8c10-5fe6993ea61f

+++ maps.cc
@@ -195,7 +195,7 @@ char *Maps::forwardGetRequest(Library *library, Elf_Addr offset,
   if (NOINTR(write(fds_[1], &req, sizeof(Request))) != sizeof(Request) ||
       NOINTR(read(fds_[0], &req.length, sizeof(req.length))) !=
       sizeof(req.length) ||
-      req.length == (size_t)-1 ||
+      req.length == -1 ||
       NOINTR(read(fds_[0], buf, length)) != (ssize_t)length) {
     memset(buf, 0, length);
     return NULL;
